requirements.txt
connexion < 3.0.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
flask_sqlalchemy
flask_prometheus_metrics

Flask
six
prometheus_client
Werkzeug
SQLAlchemy

python-logging-loki
pytz

opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask